+++ src/bool/mod.rs
@@ -1,5 +1,0 @@
-mod then_err;
-mod then_none;
-
-pub use then_err::ThenErr;
-pub use then_none::ThenNone;